Eval_Deploy/pyproject.toml
[tool.poetry]
name = "eval_deploy"
version = "0.1.0"
description = ""
authors = ["carltechiz <carl@opentechiz.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-storage = "^1.42.1"
pandas = "^1.3.2"
tensorflow = "^2.6.0"
wandb = "^0.12.1"
Pillow = "^8.3.2"
requests = "^2.26.0"
pymongo = "^3.12.0"
python-dotenv = "^0.19.0"
torch = "^1.9.1"
torchvision = "^0.10.1"
split-folders = "^0.4.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "21.4b0"
pytest-cov = "^2.12.1"
pytest-black = "^0.3.12"
flake8 = "^3.9.2"
wandb = "^0.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
